fix(build): extract pure vessel-field parsers to client-safe leaf; guard bunker lookup
The client EconomicsTab imported parseConsumption from tce-calculator, which
transitively pulls better-sqlite3 (via canal modules) into the client bundle →
'Module not found: fs' at next build. Extract parseLeadingNumber/parseConsumption
into lib/matching/parse-vessel-fields.ts (no server deps); tce-calculator
re-exports them for server callers; EconomicsTab imports from the leaf.

Also wrap getLatestBunkerPrice (bunker parity fix) in try/catch so persist/compute
don't throw on a missing bunker_prices table (minimal test DBs) — falls back to the
helper default, preserving list==detail parity.

Verified: next build exit 0, tsc --noEmit clean, jest 445/445 (lib/matching +
parity + economics-tab), parity audit 244/244 exact.
